requirements.txt
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.24.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
cryptography==45.0.7
defusedxml==0.7.1
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.32.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
magika==0.6.2
mammoth==1.10.0
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
networkx==3.5
numpy==2.3.2
olefile==0.47
onnxruntime==1.22.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pillow==11.3.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
xlrd==2.0.2
xlsxwriter==3.2.5
youtube-transcript-api==1.0.3
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
